refactor: data layer refactor/simplify

+++ src/main/java/io/github/charlespockert/data/dto/CompanyDto.java
@@ -1,9 +1,11 @@
 package io.github.charlespockert.data.dto;
 
+import java.math.BigDecimal;
+
 public class CompanyDto {
 	public int id;
 	public String name;
 	public boolean bankrupt;
 	public int sharesIssued;
-	public float value;
+	public BigDecimal value;
 }
